adjust screw diameters
As all the screws have the same diameter,
they should ideally require the same hole size everywhere,
but for some reason they do not.
The holes for the Micro-controller (Arduino) and the vowels need to be considerable smaller
then the rest, or else the screws will simply not "grab" any plastic,
and fall out by gravity.
This is based on my experiments when printing with the printer:
"Creality 3D  - CR10"